+++ src/components/Radio/Radio.tsx
@@ -11,16 +11,36 @@ export type RadioOptionType = {
   disabled?: boolean,
 };
 
+export type RadioColor = 'primary' | 'secondary' | 'success' | 'warning' | 'danger' | 'default';
+export type RadioSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
+
 type RadioButtonProps = {
   onChange: (value: string) => void,
   option: RadioOptionType,
-  color?: string,
-  size?: { button: string, label: string },
+  color?: RadioColor,
+  size?: RadioSize,
   isDisabled?: boolean,
   selected?: boolean,
 };
 
-const Radio = ({ option, selected = false, onChange, color, size, isDisabled }: RadioButtonProps) => {
+const colors = {
+  'primary': 'dsr-bg-primary',
+  'secondary': 'dsr-bg-secondary',
+  'success': 'dsr-bg-green-500',
+  'danger': 'dsr-bg-red-500',
+  'warning': 'dsr-bg-yellow-400',
+  'default': 'dsr-bg-gray-500/70',
+};
+
+const sizes = {
+  'xs': { button: 'dsr-h-2 dsr-w-2', label: 'dsr-text-xs' },
+  'sm': { button: 'dsr-h-3 dsr-w-3', label: 'dsr-text-sm' },
+  'md': { button: 'dsr-h-4 dsr-w-4', label: 'dsr-text-md' },
+  'lg': { button: 'dsr-h-5 dsr-w-5', label: 'dsr-text-lg' },
+  'xl': { button: 'dsr-h-6 dsr-w-6', label: 'dsr-text-xl' },
+};
+
+const Radio = ({ option, selected = false, onChange, color = 'primary', size = 'md', isDisabled }: RadioButtonProps) => {
 
   const { isDarkTheme } = useContext(DSRContext);
 
@@ -44,15 +64,15 @@ const Radio = ({ option, selected = false, onChange, color, size, isDisabled }: 
           <span
               className={clsx([
                 'dsr-inline-flex dsr-items-center dsr-justify-center dsr-flex-shrink-0',
-                'dsr-border-none dsr-rounded-full dsr-text-white', size?.button,
+                'dsr-border-none dsr-rounded-full dsr-text-white', sizes[size]?.button,
                 selected && styled.radioButton,
-                selected ? color :
+                selected ? colors[color] :
                   isDarkTheme ?
                     'dsr-bg-white/20' :
                     'dsr-bg-gray-500/20',
               ])}
           />
-          <span className={clsx([ 'dsr-ml-2', size?.label ])}>{ option.label }</span>
+          <span className={clsx([ 'dsr-ml-2', sizes[size]?.label ])}>{ option.label }</span>
       </div>
   );
 };
