+++ src/components/Radio/RadioGroup.tsx
@@ -1,10 +1,7 @@
 import React from 'react';
 import clsx from 'clsx';
 
-import Radio, { RadioOptionType } from './Radio';
-
-type RadioColor = 'primary' | 'secondary' | 'success' | 'warning' | 'danger' | 'default';
-export type RadioSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
+import Radio, { RadioColor, RadioOptionType, RadioSize } from './Radio';
 
 type RadioGroupType = {
   onChange: (value: string) => void,
@@ -14,23 +11,6 @@ type RadioGroupType = {
   size?: RadioSize, 
   isDisabled?: boolean,
   alignment?: 'horizontal' | 'vertical',
-};
-
-const colors = {
-  'primary': 'dsr-bg-primary',
-  'secondary': 'dsr-bg-secondary',
-  'success': 'dsr-bg-green-500',
-  'danger': 'dsr-bg-red-500',
-  'warning': 'dsr-bg-yellow-400',
-  'default': 'dsr-bg-gray-500/70',
-};
-
-const sizes = {
-  'xs': { button: 'dsr-h-2 dsr-w-2', label: 'dsr-text-xs' },
-  'sm': { button: 'dsr-h-3 dsr-w-3', label: 'dsr-text-sm' },
-  'md': { button: 'dsr-h-4 dsr-w-4', label: 'dsr-text-md' },
-  'lg': { button: 'dsr-h-5 dsr-w-5', label: 'dsr-text-lg' },
-  'xl': { button: 'dsr-h-6 dsr-w-6', label: 'dsr-text-xl' },
 };
 
 const RadioGroup = ({ onChange, value, options, color = 'primary', size = 'md', isDisabled = false, alignment = 'vertical' }: RadioGroupType) => {
@@ -46,8 +26,8 @@ const RadioGroup = ({ onChange, value, options, color = 'primary', size = 'md', 
                   key={index}
                   option={option}
                   selected={value === option.value}
-                  color={colors[color]}
-                  size={sizes[size]}
+                  color={color}
+                  size={size}
                   onChange={() => onChange(option.value)}
                   isDisabled={isDisabled}
               />
